refactored to pass tests...
